Make OTAnalytics CLI be able to run headless
