+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_analysis.py
@@ -4,7 +4,7 @@ from typing import Any
 from customtkinter import CTkButton, CTkFrame, CTkLabel
 
 from OTAnalytics.application.application import OTAnalyticsApplication
-from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 
 
 class FrameAnalysis(CTkFrame):
