+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_tracks.py
@@ -5,7 +5,7 @@ from typing import Any
 from customtkinter import CTkButton, CTkFrame, CTkLabel
 
 from OTAnalytics.application.application import OTAnalyticsApplication
-from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 
 
 class FrameTracks(CTkFrame):
