+++ OTAnalytics/plugin_ui/customtkinter_gui/toplevel_sections.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from customtkinter import CTkButton, CTkEntry, CTkLabel, CTkToplevel
 
-from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 
 
 class ToplevelSections(CTkToplevel):
