+++ OTAnalytics/plugin_ui/main_application.py
@@ -1,8 +1,5 @@
 from typing import Any
 
-import customtkinter
-from customtkinter import CTk
-
 from OTAnalytics.adapter_intersect.intersect import (
     ShapelyIntersectImplementationAdapter,
 )
@@ -40,54 +37,7 @@ from OTAnalytics.plugin_ui.cli import (
     CliParseError,
     OTAnalyticsCli,
 )
-from OTAnalytics.plugin_ui.constants import PADX, STICKY
-from OTAnalytics.plugin_ui.frame_analysis import FrameAnalysis
-from OTAnalytics.plugin_ui.frame_canvas import FrameCanvas
-from OTAnalytics.plugin_ui.frame_sections import FrameSections
-from OTAnalytics.plugin_ui.frame_tracks import FrameTracks
 from OTAnalytics.plugin_video_processing.video_reader import MoviepyVideoReader
-
-
-class OTAnalyticsGui:
-    def __init__(self, application: OTAnalyticsApplication, app: CTk = CTk()) -> None:
-        self._application = application
-        self._app: CTk = app
-
-    def start(self) -> None:
-        self._show_gui()
-
-    def _show_gui(self) -> None:
-        customtkinter.set_appearance_mode("System")
-        customtkinter.set_default_color_theme("green")
-
-        self._app.title("OTAnalytics")
-
-        self._get_widgets()
-        self._place_widgets()
-        self._wire_widgets()
-        self._app.mainloop()
-
-    def _get_widgets(self) -> None:
-        self.frame_canvas = FrameCanvas(master=self._app, application=self._application)
-        self.frame_tracks = FrameTracks(master=self._app, application=self._application)
-        self.frame_sections = FrameSections(
-            master=self._app, application=self._application
-        )
-        self.frame_analysis = FrameAnalysis(
-            master=self._app, application=self._application
-        )
-
-    def _place_widgets(self) -> None:
-        PADY = 10
-        self.frame_canvas.grid(
-            row=0, column=0, rowspan=3, padx=PADX, pady=PADY, sticky=STICKY
-        )
-        self.frame_tracks.grid(row=0, column=1, padx=PADX, pady=PADY, sticky=STICKY)
-        self.frame_sections.grid(row=1, column=1, padx=PADX, pady=PADY, sticky=STICKY)
-        self.frame_analysis.grid(row=2, column=1, padx=PADX, pady=PADY, sticky=STICKY)
-
-    def _wire_widgets(self) -> None:
-        self.frame_canvas.register_at(self._application.track_view_state)
 
 
 class ApplicationStarter:
@@ -107,6 +57,8 @@ class ApplicationStarter:
         return CliArgumentParser()
 
     def start_gui(self) -> None:
+        from plugin_ui.customtkinter_gui.gui import OTAnalyticsGui
+
         application = OTAnalyticsApplication(**self.build_dependencies())
         OTAnalyticsGui(application).start()
